// packages/api/src/database/entities/product-fabric.entity.d.ts
import { Product } from './product.entity';
export declare class ProductFabric {
    id: string;
    productId: string;
    product: Product;
    fabricCode: string;
    isPrimary: boolean;
    estimatedConsumption?: number;
    finalConsumption?: number;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=product-fabric.entity.d.ts.map