// packages/api/src/database/entities/product-sketch.entity.d.ts
import { Product } from './product.entity';
export declare class ProductSketch {
    id: string;
    productId: string;
    product: Product;
    sketchFilePath: string;
    patternId?: string;
    version: number;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=product-sketch.entity.d.ts.map